Fix asyncom version to the current stable one.
It's a bit hacky, but works. The key is that databases,
is a nice library that hides a lot of the complexity of
managing itself asyncpg,

But on the other side, it's hard to plug things inside it,
and guillotina also manages asyncpg, then, the best option
it's to just adapt a bit it, to make it work :)

Also improved a bit the docs.