Basic testing infra.
The asyncom fixture is a bit ugly, because it uses two connections,
one for the main application, based on asyncpg an another one, using,
psycopg and sqlalchemy to interact with postgres.

The second one is used, to reacreate sqlalchemy schema on each test.
Later, after yielding it drops all schema.

This, usually is done on a single global connection, and without
recreating all the schema, just using a transaction, and rolling it
back after each test (asyncpg supports nested transactions)

We need to think, if we want to evolve this direction